Improve structured responses
Due to the number of various models available in Bedrock, some of the less capable models struggle with structured output without additional instruction.  Updated the instructions with a more straight-forward schema and additional instruction.  With this change, 47 out of 51 tested models are able to repeatedly generate structured output. Unfortunately some of the models just aren't cut out for structured output